Allow combining shards in anvil
Does not include the exponential growth bug from 1.4.7. Adds the kills from the shard on the right to the one on the left.

+++ src/main/java/com/whammich/sstow/util/Utils.java
@@ -38,7 +38,7 @@ public final class Utils {
         return lastResort;
     }
 
-    public static void increaseShardKillCount(ItemStack shard, short amount) {
+    public static void increaseShardKillCount(ItemStack shard, int amount) {
         if (!shard.hasTagCompound() || hasMaxedKills(shard))
             return;
